+++ src/main/java/net/imglib2/blk/copy/RangeCopier.java
@@ -29,7 +29,7 @@ interface RangeCopier< T >
 	 * @param size
 	 * 		dimensions of block to copy from src Img.
 	 */
-	void copy( final int[] srcPos, final T dest, final int[] size );
+	void copy( final long[] srcPos, final T dest, final int[] size );
 
 	/**
 	 * Return a new independent instance of this {@code RangeCopier}. This is
